extract fetcher_config_from() helper out of engine start()
